backend/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.12.3
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==3.14.0
feedparser==6.0.10
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.3
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.9.0
